Extract product selection hook
